psa_farm_interview First Versions for 2021
The first time in GitHub for the farmer interview. I am putting in the xlsx and the csvs. i am trying a version naming convention V1_2021. This is not the first version on a farm interview, last year it was a social science interview called "Field History" that never went into the database and psa_field, just site specific questions.